requirements/base.txt
django
djangorestframework-simplejwt==5.3.0
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter
django-modeltranslation
Pillow==9.5.0
# compress image
django-resized==1.0.2

# testing
factory-boy==3.2.1
model_bakery==1.11.0
pytest==7.3.1
#pytest-mock==3.11.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
#pytest-env==0.8.2
#pytest-xdist==3.3.1

# pdf certificate
weasyprint==57.2
lxml==4.9.2
qrcode==7.3.1
pdf2image==1.16.2
